# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Projection {
  id: number;
  yearLabel: string;
  sarRate: number;
  numClients: number;
  marginPercent: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectionBody {
  yearLabel: string;
  sarRate?: number;
  numClients?: number;
  marginPercent?: number;
}

export interface UpdateProjectionBody {
  yearLabel?: string;
  sarRate?: number;
  numClients?: number;
  marginPercent?: number;
}

export interface Employee {
  id: number;
  projectionId: number;
  name: string;
  title: string;
  country: string;
  salarySar: number;
  monthsFte: number;
  ctc: number;
  totalYearlyCost: number;
  createdAt: string;
}

export interface CreateEmployeeBody {
  name: string;
  title: string;
  country: string;
  salarySar: number;
  monthsFte?: number;
}

export interface UpdateEmployeeBody {
  name?: string;
  title?: string;
  country?: string;
  salarySar?: number;
  monthsFte?: number;
}

export interface Subscription {
  id: number;
  projectionId: number;
  name: string;
  currency: string;
  originalPrice: number;
  sarEquivalent: number;
  yearlySar: number;
  monthlySar: number;
  createdAt: string;
}

export interface CreateSubscriptionBody {
  name: string;
  currency: string;
  originalPrice: number;
}

export interface UpdateSubscriptionBody {
  name?: string;
  currency?: string;
  originalPrice?: number;
}

export interface SalesSupportResource {
  id: number;
  projectionId: number;
  title: string;
  country: string;
  salarySar: number;
  months: number;
  marginPercent: number;
  ctc: number;
  totalSalaryCost: number;
  createdAt: string;
}

export interface CreateSalesSupportBody {
  title: string;
  country: string;
  salarySar: number;
  months?: number;
  marginPercent?: number;
}

export interface UpdateSalesSupportBody {
  title?: string;
  country?: string;
  salarySar?: number;
  months?: number;
  marginPercent?: number;
}

export interface Quotation {
  id: number;
  projectionId?: number | null;
  quotationNumber: string;
  companyName: string;
  clientName: string;
  date: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface QuotationLineItem {
  id: number;
  quotationId: number;
  sortOrder: number;
  description: string;
  quantity: number;
  unit: string;
  priceMonthly: number;
  totalMonths: number;
}

export interface QuotationWithLineItems {
  id: number;
  projectionId?: number | null;
  quotationNumber: string;
  companyName: string;
  clientName: string;
  date: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  lineItems: QuotationLineItem[];
}

export interface CreateQuotationBody {
  projectionId?: number | null;
  quotationNumber: string;
  companyName: string;
  clientName: string;
  date: string;
  status?: string;
}

export interface UpdateQuotationBody {
  quotationNumber?: string;
  companyName?: string;
  clientName?: string;
  date?: string;
  status?: string;
}

export interface CreateLineItemBody {
  sortOrder?: number;
  description: string;
  quantity?: number;
  unit?: string;
  priceMonthly?: number;
  totalMonths?: number;
}

export interface UpdateLineItemBody {
  sortOrder?: number;
  description?: string;
  quantity?: number;
  unit?: string;
  priceMonthly?: number;
  totalMonths?: number;
}

export interface Currency {
  id: number;
  code: string;
  name: string;
  symbol: string;
  rateToSar: number;
  isActive: boolean;
  isBase: boolean;
  updatedAt: string;
}

export interface CreateCurrencyBody {
  code: string;
  name: string;
  symbol: string;
  rateToSar: number;
  isActive?: boolean;
  isBase?: boolean;
}

export interface UpdateCurrencyBody {
  code?: string;
  name?: string;
  symbol?: string;
  rateToSar?: number;
  isActive?: boolean;
}

export interface CtcRule {
  id: number;
  countryName: string;
  countryCode: string;
  ctcMultiplier: number;
  notes?: string | null;
  isActive: boolean;
}

export interface CreateCtcRuleBody {
  countryName: string;
  countryCode: string;
  ctcMultiplier: number;
  notes?: string;
  isActive?: boolean;
}

export interface UpdateCtcRuleBody {
  countryName?: string;
  countryCode?: string;
  ctcMultiplier?: number;
  notes?: string;
  isActive?: boolean;
}

export interface SystemSettings {
  id: number;
  defaultMargin: number;
  defaultNumClients: number;
  vatRate: number;
  quotationPrefix: string;
  companyName: string;
  companyLogoUrl?: string | null;
  baseCurrencyCode: string;
  updatedAt: string;
}

export interface UpdateSystemSettingsBody {
  defaultMargin?: number;
  defaultNumClients?: number;
  vatRate?: number;
  quotationPrefix?: string;
  companyName?: string;
  companyLogoUrl?: string;
  baseCurrencyCode?: string;
}

export interface ChatMessageRequest {
  message: string;
  projectionId?: number | null;
}

export type ChatMessageResponseActionsTakenItem = {
  action: string;
  details: string;
};

export interface ChatMessageResponse {
  reply: string;
  actionsTaken: ChatMessageResponseActionsTakenItem[];
}

export interface ProjectionSummary {
  projection: Projection;
  totalDeptCostYearly: number;
  costPerClientYearly: number;
  costPerClientMonthly: number;
  totalOverheadMonthly: number;
  totalOverheadYearly: number;
  totalMonthlyCostPerClient: number;
  sellingPriceWithoutVat: number;
  marginSarMonthly: number;
  marginSarYearly: number;
  sellingPriceWithVatMonthly: number;
  sellingPriceWithVatYearly: number;
  salesSupportTotalCost: number;
  salesSupportSellingPrice: number;
  employeeCount: number;
  subscriptionCount: number;
  salesSupportCount: number;
}

export interface DashboardSummary {
  totalProjections: number;
  totalQuotations: number;
  recentProjection?: ProjectionSummary;
  recentQuotations: Quotation[];
}