# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Department Projection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projections:
    get:
      operationId: listProjections
      tags: [projections]
      summary: List all projections
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Projection"
    post:
      operationId: createProjection
      tags: [projections]
      summary: Create a projection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projection"
  /projections/{id}:
    get:
      operationId: getProjection
      tags: [projections]
      summary: Get a projection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projection"
    put:
      operationId: updateProjection
      tags: [projections]
      summary: Update projection settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projection"
  /projections/{id}/summary:
    get:
      operationId: getProjectionSummary
      tags: [dashboard]
      summary: Full computed projection summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionSummary"
  /projections/{projectionId}/employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Add employee
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /projections/{projectionId}/employees/{id}:
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Remove employee
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/currencies:
    get:
      operationId: listCurrencies
      tags: [admin]
      summary: List currencies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Currency"
    post:
      operationId: createCurrency
      tags: [admin]
      summary: Add currency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCurrencyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
  /admin/currencies/{id}:
    put:
      operationId: updateCurrency
      tags: [admin]
      summary: Update currency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCurrencyBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
    delete:
      operationId: deleteCurrency
      tags: [admin]
      summary: Delete currency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/ctc-rules:
    get:
      operationId: listCtcRules
      tags: [admin]
      summary: List CTC rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CtcRule"
    post:
      operationId: createCtcRule
      tags: [admin]
      summary: Add CTC rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCtcRuleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtcRule"
  /admin/ctc-rules/{id}:
    put:
      operationId: updateCtcRule
      tags: [admin]
      summary: Update CTC rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCtcRuleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtcRule"
    delete:
      operationId: deleteCtcRule
      tags: [admin]
      summary: Delete CTC rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/settings:
    get:
      operationId: getSystemSettings
      tags: [admin]
      summary: Get system settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    put:
      operationId: updateSystemSettings
      tags: [admin]
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemSettingsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
...
components:
  schemas:
...
    Projection:
      type: object
      properties:
        id:
          type: integer
        yearLabel:
          type: string
        sarRate:
          type: number
        numClients:
          type: integer
        marginPercent:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, yearLabel, sarRate, numClients, marginPercent, createdAt, updatedAt]
    CreateProjectionBody:
      type: object
      properties:
        yearLabel:
          type: string
        sarRate:
          type: number
        numClients:
          type: integer
        marginPercent:
          type: number
      required: [yearLabel]
    UpdateProjectionBody:
      type: object
      properties:
        yearLabel:
          type: string
        sarRate:
          type: number
        numClients:
          type: integer
        marginPercent:
          type: number
    Employee:
      type: object
      properties:
        id:
          type: integer
        projectionId:
          type: integer
        name:
          type: string
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        monthsFte:
          type: integer
        ctc:
          type: number
        totalYearlyCost:
          type: number
        createdAt:
          type: string
      required: [id, projectionId, name, title, country, salarySar, monthsFte, ctc, totalYearlyCost, createdAt]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        monthsFte:
          type: integer
      required: [name, title, country, salarySar]
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        monthsFte:
          type: integer
...
    Currency:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
        rateToSar:
          type: number
        isActive:
          type: boolean
        isBase:
          type: boolean
        updatedAt:
          type: string
      required: [id, code, name, symbol, rateToSar, isActive, isBase, updatedAt]
    CreateCurrencyBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
        rateToSar:
          type: number
        isActive:
          type: boolean
        isBase:
          type: boolean
      required: [code, name, symbol, rateToSar]
    UpdateCurrencyBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
        rateToSar:
          type: number
        isActive:
          type: boolean
    CtcRule:
      type: object
      properties:
        id:
          type: integer
        countryName:
          type: string
        countryCode:
          type: string
        ctcMultiplier:
          type: number
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, countryName, countryCode, ctcMultiplier, isActive]
    CreateCtcRuleBody:
      type: object
      properties:
        countryName:
          type: string
        countryCode:
          type: string
        ctcMultiplier:
          type: number
        notes:
          type: string
        isActive:
          type: boolean
      required: [countryName, countryCode, ctcMultiplier]
    UpdateCtcRuleBody:
      type: object
      properties:
        countryName:
          type: string
        countryCode:
          type: string
        ctcMultiplier:
          type: number
        notes:
          type: string
        isActive:
          type: boolean
    SystemSettings:
      type: object
      properties:
        id:
          type: integer
        defaultMargin:
          type: number
        defaultNumClients:
          type: integer
        vatRate:
          type: number
        quotationPrefix:
          type: string
        companyName:
          type: string
        companyLogoUrl:
          type: string
          nullable: true
        baseCurrencyCode:
          type: string
        updatedAt:
          type: string
      required: [id, defaultMargin, defaultNumClients, vatRate, quotationPrefix, companyName, baseCurrencyCode, updatedAt]
    UpdateSystemSettingsBody:
      type: object
      properties:
        defaultMargin:
          type: number
        defaultNumClients:
          type: integer
        vatRate:
          type: number
        quotationPrefix:
          type: string
        companyName:
          type: string
        companyLogoUrl:
          type: string
        baseCurrencyCode:
          type: string
...
    ProjectionSummary:
      type: object
      properties:
        projection:
          $ref: "#/components/schemas/Projection"
        totalDeptCostYearly:
          type: number
        costPerClientYearly:
          type: number
        costPerClientMonthly:
          type: number
        totalOverheadMonthly:
          type: number
        totalOverheadYearly:
          type: number
        totalMonthlyCostPerClient:
          type: number
        sellingPriceWithoutVat:
          type: number
        marginSarMonthly:
          type: number
        marginSarYearly:
          type: number
        sellingPriceWithVatMonthly:
          type: number
        sellingPriceWithVatYearly:
          type: number
        salesSupportTotalCost:
          type: number
        salesSupportSellingPrice:
          type: number
        employeeCount:
          type: integer
        subscriptionCount:
          type: integer
        salesSupportCount:
          type: integer
      required: [projection, totalDeptCostYearly, costPerClientYearly, costPerClientMonthly, totalOverheadMonthly, totalOverheadYearly, totalMonthlyCostPerClient, sellingPriceWithoutVat, marginSarMonthly, marginSarYearly, sellingPriceWithVatMonthly, sellingPriceWithVatYearly, salesSupportTotalCost, salesSupportSellingPrice, employeeCount, subscriptionCount, salesSupportCount]
...