# lib/api-zod/src/generated/types/createCtcRuleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCtcRuleBody {
  countryName: string;
  countryCode: string;
  ctcMultiplier: number;
  notes?: string;
  isActive?: boolean;
}