# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  projectionId: number;
  name: string;
  title: string;
  country: string;
  salarySar: number;
  monthsFte: number;
  ctc: number;
  totalYearlyCost: number;
  createdAt: string;
}