# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEmployeeBody {
  name?: string;
  title?: string;
  country?: string;
  salarySar?: number;
  monthsFte?: number;
}